Bug 9414 - Fix sample letters for German and Norwegian
Some letters were missing or not translated in the German and
Norwegian data for the webinstaller. This patch adds them.

German
- RENEWAL (present, but not translated)
- OPAC_REG_VERIFY (missing)

Norwegian
- RENEWAL (present, but not translated)
- ISSUESLIP (missing)
- ISSUEQSLIP (missing)
- RESERVESLIP (missing)
- TRANSFERSLIP (missing)
- OPAC_REG_VERIFY (missing)

To test:
- Install Norwegian and German* translations
- Run through the web installer in German, making sure that
  the sample letters are selected
- Check that there are no errors in the web installer
- Check <intranet>/cgi-bin/koha/tools/letter.pl - that the
  letters look OK, and are translated
- Repeat the process for Norwegian

* = I had some problems with installing the German
translations that are in master, but now I can't reproduce
it. If you run into a problem where only some of the
subdirectories of koha-tmpl/intranet-tmpl/prog/de-DE/modules/
are cerated, try downloading and installing the 3.10 .po
files from Pootle.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>