Bug 9877 - Use DataTables on funds administration page
Replace the tablesorter plugin with the DataTables plugin on the funds
administration page.

To test, open the funds administration page (Administration ->
Funds). Confirm that table sorting, paging, and searching works
correctly.

Note that the table is sorted only when hierarchical display of parent
and child funds is not enabled. When the hierarchicial display is
enabled the table sorter is not triggered.

Signed-off-by: Pierre Angot <tredok.pierre@gmail.com>

Post-QA revision adds handling for empty table: "No funds" row must have
same number of cells as the rest of the table (including hidden cell);

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>