Bug 9850: Update Patron Help Files for 3.12
This patch updates all the help files in the members folder. It also
adds a missing help file for approving patron account changes.

To Test:

Visit the help on every patron page and make sure the text and
manual links are all right.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Fixed 2 obvious typos mentioned on the bug report.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>