Bug 9947 - Creating item for received serial prefills enumchron subfield
In serial receiving :
When creating an item for an issue, a javascript prefilles the $h subfield with issue number.
This is for MARC21, in UNIMARC serial number is on $v.

This patch replaces hard-coded value with the subfield mapped with "items.enumchron".

Test plan :
- Set "Koha to MARC mapping" of "items.enumchron" on a subfield other than $h, ie $z
- Go to a serial subscription detail with "Serial receipt creates an item record"
- Click on "Receive"
- On an issue, click on "Click to add item"
=> Check that issue number is copied into selected subfield, ie $z
You may test that with no mapping of "items.enumchron" there is no javascript error

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Nice catch! Works as advertised.
All tests pass after fixing tabulation characters in serials-edit.tt.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>