Bug 9970 - Add datepicker default configuration to OPAC patron entry form
The datepicker on the OPAC patron entry/update form
(opac-memberentry.pl) doesn't show a calendar icon. This is because
although there is in-page JS to trigger the widget, the page doesn't
include the file which sets some widget defaults, including the path to
the calendar icon and the translatable strings for calendar labels.

This patch simply adds calendar.inc to the template.

To test, load the patron entry/update form (click "personal details" for
an existing patron or a start new patron registration). The date of
birth entry field should show a calendar icon.

Signed-off-by: David Cook <dcook@prosentient.com.au>

Test plan works as described. Looks much better.

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>