Bug 8061: Follow Up (Opac-Prog) - enhancement to D Ruth's patch
In this Patch I added some changes to the rtl.css file
I made some changes in the file opac-bottom.inc and add css
classes to the lang list and I define the class value in
the opac.css
I add If statment to the opac-detail.tt to chose the rtl.css
when rtl lang.

http://bugs.koha-community.org/show_bug.cgi?id=8061

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Reformatted commit message and added bug number.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Tested with po files for Arabic from 3.8.
All tests and QA script pass.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>