Bug 10028: fix how ModReceiveOrder() calls NewOrder()
In C4::Acquisition::ModReceiveOrder, a call to NewOrder is badly used.

NewOrder returns ($basketno, $ordernumber) but in ModReceiveOrder the
ordernumber is got with
  my $ordernumber = NewOrder( $args );

It works because:
sub t{
    return ("a", "b");
}
my $a = t();
say $a;

Will display 'b'.

But it is not really clear.

Test plan:
Check that there is no regression for partial receives.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>