Bug 10052: Db revision for intranetstylesheet
Report 10052 makes the intranetstylesheet prefs behave like the opac ones.
This patch contains an optional db revision: If the pref intranetstylesheet is
not blank, and does not start with http, https or /, then prefix it with a
slash to make it an absolute path.

Test plan:
Before applying this patch:
1) Put intranet-tmpl/prog/en/css/staff-global.css into intranetstylesheet.
This works in the current staff client, but will be interpreted as relative as
soon as both patches for this report have been applied.
Apply the patch. And run: /cgi-bin/koha/installer/install.pl?step=3&op=updatestructure
Check the webinstaller warning and the pref value. There should be a slash
in front now. The css should still work.

Continue after applying this patch with:
2) Clear the intranetstylesheet pref. Run updatestructure again. No message: OK
3) Put /intranet-tmpl/prog/en/css/staff-global.css in intranetstylesheet.
Run updatestructure again. No message: OK
4) Prefix intranetstylesheet with http://[your-server-here]. Test again.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Comments on last patch.
Signed-off-by: Galen Charlton <gmc@esilibrary.com>