Bug 10017:  Follow up OPAC-CCSR-Right-to-left-css
This follow up patch will delete the "-moz-box-sizing"
property that I was add when I'm doing some tests on
the rtl css file and  I forgot to delete it

many thanks David Cook for your notes deleting that property will not effect  on the results

Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>

Comment on last patch.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Comment on last patch.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>