Bug 9266 - Untranslated strings in tags moderation
There are a couple of English strings in tags moderation JavaScript
which are not properly enclosed in the _() function enabling
translation. As a result, clicking a translated 'Approve' or
'Reject' button in the table of tags would change the button's
neighbor's label to the English string. This patch
corrects it.

To test, apply the patch and regenerate a translation to test with. Go
to Tools -> Tags and find a list of tags to moderate (pending, approved,
or rejected). When you click on an 'Approve' or 'Reject' tag in the
table of tags, the buttons in each Approve/Reject pair should retain
their correct translated label. For instance, looking at approved tags
in French:

[ Approuvé ] [ Rejeter ]

Clicking the "Rejetee" button should change the state to:

[ Approuver ] [ Rejetée ]

Clicking "Approuver" should return the buttons to the first state.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Mason James <mtj@kohaaloha.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>