Bug 10083 - In Transit string doesn't get translated in the staff interface
For some reason this text inside a <td> isn't getting picked up by the
translator unless it's inside a <span>. This patch adds the span and
tidies up the text.

To test, apply the patch and "translate update" a translation. Searching
the po file for "In transit from" should return a result from detail.pl.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Template only change fixing a translation problem.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>