Bug 10191: Close 2 option tags in unimarc_field_115a.tt and unimarc_field_116.tt
This patch close 2 option tags in unimarc_field_115a.tt and unimarc_field_116.tt.

To test :
- in a UNIMARC Koha instance, edit a record
- open the plugin for 115a subfield
- check the "Form of release - visual projection, motion picture" option
  list is correct, especially the value "other film type"
- open the plugin for 116 field
- check the "Technique (prints) 2" option list is correct, especially
  the value "lithography"

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

I did not test this in Koha, but it's a simple change of one character
which I'm comfortable with signing off on.

I don't know if this counts as a string change or not, since the string
was previously mixed with a malformed tag.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Passes tests and fixes an obvious problem in the HTML.
I checked the po files - currently the strings appear as:
"other film type/option> %s "

So in theory it's possible to translate them correctly even
if the standard templates now are not correct.
Probably has to wait for 3.14.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>