Bug 10166: Increment holdable counter
The $anyholdable variable was set to 0 or 1. However, as it is set in
a loop, and future changes to the opac-reserve.pl script may require
knowing how many items the patron is going to place a hold on, it makes
more sense to treat $anyholdable as a counter. This follow up turns it
into one.

Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>