Bug 6329 - patron categories should show which fields are required
When entering a new patron category, required fields are not indicated
on the entry form. This patch adds the "required" class to the relevant
labels.

To test, add or edit a patron category. Confirm that the required fields
(category code, category type, description, and enrollment period) are
indicated as being required with a red label.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works nicely.
Template changes only, but no string changes.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>